requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
hnswlib
numpy
pypdf
python-docx
tqdm
requests
pydantic
